feat: refactor component imports and add organization management features
