+++ apps/assassin/src/components/org/OrgSettingsPageClient.tsx
@@ -10,9 +10,9 @@ import { Separator } from "@/components/ui/separator";
 import { useRealtimeOrgSync } from "@/features/org/hooks/useRealtimeOrgSync";
 import { useOrgBySlug } from "@/features/org/queries/useOrgBySlug";
 import { useOrgRole } from "@libs/org/OrgProvider";
-import { useUpdateOrg } from "../mutations/useUpdateOrg";
-import { DeleteOrgButton } from "./DeleteOrgButton";
-import { LeaveOrgButton } from "./LeaveOrgButton";
+import { useUpdateOrg } from "@/features/org/mutations/useUpdateOrg";
+import { DeleteOrgButton } from "@/components/org/DeleteOrgButton";
+import { LeaveOrgButton } from "@/components/org/LeaveOrgButton";
 
 interface OrgSettingsPageClientProps {
   orgSlug: string;
